% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{Function to load or install package}
\usage{
ipak(pkg)
}
\arguments{
\item{pkgs}{packages to load}
}
\description{
This function will load packages and install them if missing. 
It can handle a vector of package names as well.
}
\examples{
ipak()
}
\keyword{Load/Install}
\keyword{Packages}
