% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_csvs.R
\name{read_multiple_files}
\alias{read_multiple_files}
\title{Function to read multiple csvs}
\usage{
read_multiple_files(file_lists = NULL, pattern = ".csv")
}
\arguments{
\item{file_lists}{List of files to read into environment. Defaults to csvs in directory.}
}
\description{
This function will load all csvs (or file pattern ending) in path, or based on 
file lists provided.
}
\examples{
read_multiple_files()
}
\keyword{Read}
\keyword{files}
\keyword{multiple}
