% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{rotate_x_axis}
\alias{rotate_x_axis}
\title{Rotate's X axis}
\usage{
rotate_x_axis(ang)
}
\arguments{
\item{angle}{Angle at which text should be rotated}
}
\description{
This function just loads the theme specification for ggplot2 to rotate the x-axis
}
\examples{
rotate_x_xis()
}
\keyword{Rotate}
\keyword{axis}
\keyword{label}
