% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{rotate_y_axis}
\alias{rotate_y_axis}
\title{Rotate's Y axis}
\usage{
rotate_y_axis(ang)
}
\arguments{
\item{angle}{Angle at which text should be rotated}
}
\description{
This function just loads the theme specification for ggplot2 to rotate the y-axis
}
\examples{
rotate_y_xis()
}
\keyword{Rotate}
\keyword{axis}
\keyword{label}
